[수정] 구조 개선
- 기능이 추가될수록 HttpRequestHandler에 함수가 추가되면 책임이 너무 과해지는 것 같아 각 함수를 분리하였다.
- 각각의 이름을 Controller로 두고, RequestHandler는 각 class를 받아와 handle하도록 하였다.
- Response를 RequestHandler에서 render 했던 것을 별도의 class 를 만들어 분리하였다.
