+++ src/test/java/webserver/HttpRequestParserTest.java
@@ -1,9 +1,10 @@
 package webserver;
 
-import org.assertj.core.util.Maps;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.assertj.core.api.SoftAssertions;
+import webserver.http.HttpRequest;
+import webserver.http.HttpRequestParser;
 
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
